ai-services/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
motor>=3.3.2
pymongo>=4.6.0
python-dotenv>=1.0.0
python-multipart==0.0.6
httpx==0.25.2
aiofiles==23.2.1

# ML/Data Science Dependencies (Python 3.12 compatible)
numpy>=1.26.0
scikit-learn>=1.4.0
xgboost>=2.0.3
scipy>=1.12.0
pandas>=2.1.4
